Condenses code into fewer files
